+++ src/routes/+page.js
@@ -1,5 +1,1 @@
-import { redirect } from '@sveltejs/kit';
-
-export function load() {
-	throw redirect(307, '/today');
-}
+export const prerender = true;