+++ svelte.config.js
@@ -5,7 +5,7 @@ const config = {
 		adapter: adapter({
 			pages: 'build',
 			assets: 'build',
-			fallback: 'index.html',
+			fallback: '200.html',
 			strict: false
 		})
 	}